requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==20.1.0
